+++ components/parts/Header.tsx
@@ -1,0 +1,11 @@
+const Header = () => {
+    return (
+        <header>
+            <h1></h1>
+        </header>
+    );
+}
+
+export default Header;
+
+
